Page Layout
Added three divs which act as the left, right, and center of the bottom
of the page, for the old footer controls. Now that the chat div locks
automatically, there will only need to be volume controls on the right
side, clearing up a lot of space.

Added auto focus to the textarea on page load.