Hard message length limits and retry
Messages are now dropped if their encoded length exceeds 1024
characters. Dropped messages result in a 400 status code to the client.

If a message is dropped or otherwise fails (error code in 4xx or 5xx),
the client will restore the failed message to the input field to allow
the user to edit and retry (if they haven't typed anything else).