Message counter fix for updated chat.cgi
Previously the message counter assumed that the number of messages
received included an extraneous empty message always appended to the end
of the list of messages. That empty message has been removed, and so the
message counter was off by one.

The error has been corrected, and the title bar should now always
accurately reflect the number of new messages not seen.